requirements.txt
feedparser
requests
pyyaml
python-dotenv
google-generativeai
beautifulsoup4
fastapi
uvicorn
pydantic
duckduckgo-search
